backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic==2.11.7
httpx==0.28.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.9.0

chromadb==0.4.18
sentence-transformers==2.2.2
PyPDF2==3.0.1